Moved widget tree out of camera home file and into another file
